BUG: change registration type traits to handle unsigned voxel types (gets around the root problem: ITK pyramid filter does not handle unsigned types)
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=7142
git-svn-id: http://svn.slicer.org/Slicer4/trunk@7142 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/EMSegment/Registration/vtkRegistratorTypeTraits.h
@@ -26,7 +26,7 @@ template<>
 class RegistrationVoxelTypeTraits<unsigned char>
 {
 public:
-  typedef unsigned short RegistrationVoxelType;
+  typedef short RegistrationVoxelType;
 };
 
 template<>
@@ -47,7 +47,7 @@ template<>
 class RegistrationVoxelTypeTraits<unsigned short>
 {
 public:
-  typedef unsigned short RegistrationVoxelType;
+  typedef int RegistrationVoxelType;
 };
 
 template<>
